Move mapping export code into jadx-gui (for now)
