+++ jadx-core/build.gradle
@@ -4,10 +4,6 @@ plugins {
 
 dependencies {
 	api(project(':jadx-plugins:jadx-plugins-api'))
-	// TODO: Switch back to upstream once this PR gets merged:
-	//   https://github.com/FabricMC/mapping-io/pull/19
-	// api 'net.fabricmc:mapping-io:0.3.0'
-	api files('libs/mapping-io-0.4.0-SNAPSHOT.jar')
 
 	implementation 'com.google.code.gson:gson:2.9.0'
 	implementation 'com.android.tools.build:aapt2-proto:4.2.1-7147631'
