+++ jadx-gui/build.gradle
@@ -29,6 +29,11 @@ dependencies {
 	implementation 'com.android.tools.build:apksig:4.2.1'
 	implementation 'io.github.hqktech:jdwp:1.0'
 
+	// TODO: Switch back to upstream once this PR gets merged:
+	//   https://github.com/FabricMC/mapping-io/pull/19
+	// implementation 'net.fabricmc:mapping-io:0.3.0'
+	implementation files('libs/mapping-io-0.4.0-SNAPSHOT.jar')
+
 	testImplementation project(":jadx-core").sourceSets.test.output
 }
